+++ Editor/Providers/SceneObjectsProvider.cs
@@ -2,7 +2,9 @@ using JetBrains.Annotations;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using UnityEditor;
+using UnityEditor.Experimental.SceneManagement;
 using UnityEditor.ShortcutManagement;
 using UnityEngine;
 
@@ -81,7 +83,7 @@ namespace Unity.QuickSearch
                             }
 
                             var gameObjectId = go.GetInstanceID().ToString();
-                            var item = provider.CreateItem(gameObjectId, ~(int)score, $"{go.name} ({gameObjectId})", null, null, null);
+                            var item = provider.CreateItem(gameObjectId, ~(int)score, null, null, null, null);
                             item.customDescriptionFormatter = useFuzzySearch;
                             items.Add(item);
                             if (++addedCount >= 200)
@@ -89,23 +91,49 @@ namespace Unity.QuickSearch
                         }
                     };
 
-                    fetchDescription = (item, context) =>
+                    fetchLabel = (item, context) =>
                     {
+                        if (item.label != null)
+                            return item.label;
+
                         var go = ObjectFromItem(item);
-                        var sq = CleanString(context.searchQuery.ToLowerInvariant());
-
-                        item.description = go.transform.GetPath() + " (" + item.score + ")";
-
-                        const int maxCharCount = 105;
-                        if (item.description.Length > maxCharCount)
-                            item.description = "..." + item.description.Substring(item.description.Length - maxCharCount);
+                        if (!go)
+                            return item.id;
+                        item.label = $"{go.transform.GetPath()} ({item.id})";
 
                         if (item.customDescriptionFormatter)
                         {
                             long score = 1;
                             List<int> matches = new List<int>();
+                            var sq = CleanString(context.searchQuery.ToLowerInvariant());
+                            if (FuzzySearch.FuzzyMatch(sq, CleanString(item.label), ref score, matches))
+                                item.label = RichTextFormatter.FormatSuggestionTitle(item.label, matches, FuzzySearch.HighlightColorTag, FuzzySearch.HighlightColorTagSpecial);
+                        }
+
+                        return item.label;
+                    };
+
+                    fetchDescription = (item, context) =>
+                    {
+                        var go = ObjectFromItem(item);
+
+                        item.description = GetHierarchyPath(go) + go.transform.GetPath() + " (" + item.score + ")";
+
+                        const int maxCharCount = 105;
+                        if (item.description.Length > maxCharCount)
+                        {
+                            item.description = item.description.Replace("<b>", "").Replace("</b>", "");
+                            int cutPos = item.description.Length - maxCharCount;
+                            item.description = "..." + item.description.Substring(cutPos);
+                        }
+
+                        if (item.customDescriptionFormatter)
+                        {
+                            long score = 1;
+                            List<int> matches = new List<int>();
+                            var sq = CleanString(context.searchQuery.ToLowerInvariant());
                             if (FuzzySearch.FuzzyMatch(sq, CleanString(item.description), ref score, matches))
-                                item.description = RichTextFormatter.FormatSuggestionTitle(item.description, matches, "<color=#FF6100>", "<color=#FF6100>");
+                                item.description = RichTextFormatter.FormatSuggestionTitle(item.description, matches, FuzzySearch.HighlightColorTag, FuzzySearch.HighlightColorTagSpecial);
                         }
                         return item.description;
                     };
@@ -118,6 +146,17 @@ namespace Unity.QuickSearch
                         var obj = ObjectFromItem(item);
                         if (obj != null)
                         {
+                            var assetPath = GetHierarchyAssetPath(obj, true);
+                            if (!String.IsNullOrEmpty(assetPath))
+                            {
+                                item.thumbnail = AssetPreview.GetAssetPreview(obj);
+                                if (item.thumbnail)
+                                    return item.thumbnail;
+                                item.thumbnail = Utils.GetAssetThumbnailFromPath(assetPath, true);
+                                if (item.thumbnail)
+                                    return item.thumbnail;
+                            }
+
                             item.thumbnail = PrefabUtility.GetIconForGameObject(obj);
                             if (item.thumbnail)
                                 return item.thumbnail;
@@ -177,13 +216,33 @@ namespace Unity.QuickSearch
                         {
                             var pingedObject = PingItem(item);
                             if (pingedObject != null)
+                                FrameObject(pingedObject);
+                        }
+                    },
+
+                    new SearchAction(type, "open", null, "Open containing asset...")
+                    {
+                        handler = (item, context) =>
+                        {
+                            var pingedObject = PingItem(item);
+                            if (pingedObject != null)
                             {
-                                Selection.activeGameObject = pingedObject as GameObject ?? Selection.activeGameObject;
-                                SceneView.lastActiveSceneView.FrameSelected();
+                                var go = pingedObject as GameObject;
+                                var assetPath = GetHierarchyAssetPath(go);
+                                if (!String.IsNullOrEmpty(assetPath))
+                                    Utils.FrameAssetFromPath(assetPath);
+                                else
+                                    FrameObject(go);
                             }
                         }
                     }
                 };
+            }
+
+            private static void FrameObject(object obj)
+            {
+                Selection.activeGameObject = obj as GameObject ?? Selection.activeGameObject;
+                SceneView.lastActiveSceneView.FrameSelected();
             }
 
             private static GameObject ObjectFromItem(SearchItem item)
@@ -198,6 +257,107 @@ namespace Unity.QuickSearch
                 if (tform.parent == null)
                     return "/" + tform.name;
                 return tform.parent.GetPath() + "/" + tform.name;
+            }
+
+            private static readonly Stack<StringBuilder> _SbPool = new Stack<StringBuilder>();
+            public static string GetHierarchyPath(GameObject gameObject, bool includeScene = true, [CanBeNull] Component component = null)
+            {
+                if (gameObject == null)
+                    return String.Empty;
+
+                StringBuilder sb;
+                if (_SbPool.Count > 0)
+                {
+                    sb = _SbPool.Pop();
+                    sb.Clear();
+                }
+                else
+                {
+                    sb = new StringBuilder(200);
+                }
+
+                try
+                {
+
+                    bool isPrefab;
+                    #if UNITY_2018_3_OR_NEWER
+                    isPrefab = UnityEditor.PrefabUtility.GetPrefabAssetType(gameObject.gameObject) != UnityEditor.PrefabAssetType.NotAPrefab;
+                    #else
+                    isPrefab = UnityEditor.PrefabUtility.GetPrefabType(o) == UnityEditor.PrefabType.Prefab;
+                    #endif
+
+                    var assetPath = string.Empty;
+                    if (isPrefab)
+                    {
+                        #if UNITY_2018_3_OR_NEWER
+                        assetPath = PrefabUtility.GetPrefabAssetPathOfNearestInstanceRoot(gameObject);
+                        #else
+                        assetPath = AssetDatabase.GetAssetPath(gameObject);
+                        #endif
+                        sb.Append("<b>" + assetPath + "</b>");
+                    }
+                    else
+                    {
+                        if (includeScene)
+                        {
+                            var sceneName = gameObject.scene.name;
+                            if (sceneName == string.Empty)
+                            {
+                                #if UNITY_2018_3_OR_NEWER
+                                var prefabStage = PrefabStageUtility.GetPrefabStage(gameObject);
+                                if (prefabStage != null)
+                                {
+                                    sceneName = "Prefab Stage";
+                                }
+                                else
+                                #endif
+                                {
+                                    sceneName = "Unsaved Scene";
+                                }
+                            }
+
+                            sb.Append("<b>" + sceneName + "</b>");
+                        }
+                    }
+
+                    var path = sb.ToString();
+                    sb.Clear();
+                    return path;
+                }
+                finally
+                {
+                    _SbPool.Push(sb);
+                }
+
+            }
+
+            public static string GetHierarchyAssetPath(GameObject gameObject, bool prefabOnly = false)
+            {
+                if (gameObject == null)
+                    return String.Empty;
+
+                bool isPrefab;
+                #if UNITY_2018_3_OR_NEWER
+                isPrefab = UnityEditor.PrefabUtility.GetPrefabAssetType(gameObject.gameObject) != UnityEditor.PrefabAssetType.NotAPrefab;
+                #else
+                isPrefab = UnityEditor.PrefabUtility.GetPrefabType(o) == UnityEditor.PrefabType.Prefab;
+                #endif
+
+                var assetPath = string.Empty;
+                if (isPrefab)
+                {
+                    #if UNITY_2018_3_OR_NEWER
+                    assetPath = PrefabUtility.GetPrefabAssetPathOfNearestInstanceRoot(gameObject);
+                    #else
+                    assetPath = AssetDatabase.GetAssetPath(gameObject);
+                    #endif
+                    return assetPath;
+                }
+
+                if (prefabOnly)
+                    return null;
+
+                return gameObject.scene.path;
             }
 
             #if UNITY_2019_1_OR_NEWER